Fix sub region detection in nested if-else blocks
The compiler does an optimization to avoid pointing a goto at a goto so the
goto from the inner if body goes directly to the end of the **outer** else
block. This change attempts to detect that and end the nested region at the
alternate goto at the end of the outer if body.