Fix while loops immediately after if statements
Some extra processing is needed since the goto that normally
starts an if statement is optimized out and instead the target of the
if statement is changed to be the start of the while condition. This both
breaks the region detection for the if statement and makes the while look
like a do-while.